Make check_integrity.py be a wrapper script to a package module
